test(core): migrate UpdateSearchTests (Wave C)
14 tests, 26 failure sites preserved 1:1.
